Inserted code to delete bool vectors used for live variable analysis.
Added isEssential bool to primitives.  This bool signals that a
primitive should never be eliminated even if it appears to be dead
because it modifies its arguments and/or has side effects.  In
addition, return and yield are marked as essential since they produce
effects of functions.  In adding new primitives, make sure that the
isEssential flag is set correctly.  This should be relatively easy.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@9976 3a8e244f-b0f2-452b-bcba-4c88e055c3ca