Add tuple destructuring for function arguments.  The definition: def foo(a, (b,(c,d))) where cond { ...
becomes:
def foo(a, _tuple_arg: _tuple) where
      cond & _tuple_arg.size == 2 &
      PRIMITIVE_IS_TUPLE(_tuple_arg(2)) & _tuple_arg(2).size == 2 {
  var b = _tuple_arg(1);
  var c = _tuple_arg(2)(1);
  var d = _tuple_arg(2)(2); ...

Once folding of && is implemented, the where clause should be changed to use it
instead of bitwise &.

Remove a .future.  Add a more involved test of this.  Add a future showing why
the where clause needs to use short-circuiting && instead of bitwise.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10016 3a8e244f-b0f2-452b-bcba-4c88e055c3ca