More garbage collection work.  Do a breadth-first pointer traversal starting with the root-set.  To do this, generate one array per class type containing the offsets of the pointer fields in the C struct representing the class, as well as a cid2offsets function which returns pointers to these arrays.
Also, add pointers that are at module scope to the root-set so they can be
collected.  I've not yet added pointers that are in module scope records to
the root-set, nor does the pointer traversal find pointers in records in
classes.  It shouldn't be a big change to scan records in these situations
though.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10034 3a8e244f-b0f2-452b-bcba-4c88e055c3ca