Sorted through futures, updating .future files and classifying as appropriate.
Expanded on Steve's future classification scheme by
having the first line of each future follow something
like the following template:

  label: description (owner)

where label is one of the following six choices:

  bug: the test shows a bug in the implementation
  error message: the test needs a compiler error msg
                 (or, more often, a better one)
  feature request: requests a new feature for the
                   language or implementation
  runtime error: the test needs a runtime error msg
                 (or a better one)
  semantic: the test raises questions about the
            language which probably need to be
            addressed before fixing the future
  unimplemented feature: the test exercises a feature
                         that we've never implemented

description is a brief phrasing of the problem (trying
to reuse existing descriptions whenever possible), and
owner is the developer who currently owns the future
(if applicable).

The goal of using these labels is to reduce the amount
of work required to convert the list of futures into a
STATUS file.

In some cases, putting a test into one of these
categories was a judgement call -- particularly between
bug and unimplemented feature -- feel free to change
the classification of any of them if you disagree.

In addition to unifying the format of the .future files,
added and removed information from the future files to
reflect decisions that we've made and information we've
determined.  Didn't always preserve the previous content
in a log-style format, figuring we could rely on CVS to
do that for us.

This scan also resulted in the following changes:

* Removed a few more reference counting leak futures.

* Fixed a future by having our error message generator
  print "iterator" rather than "function" for bugs in
  iterators.  (several other changes had gotten this
  future within epsilon of working, this was just putting
  the icing on the cake).

* Reclassified the semantic futures that we'd reviewed
  as a group in spec meetings, and updated .good files
  as necessary.

* Updated some old domain indexing syntax to use .dim

* Fixed a thinko when I thought that writing multiple
  arguments was meant to put a space between them.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10093 3a8e244f-b0f2-452b-bcba-4c88e055c3ca