Fixed a bug in which functions in modules that were used would take precedence over functions in a given module for code defined at the top level of the module because the use was in the block of the init function rather than the block of the module.
Moved functions that conflict with standard functions in the tests

  distributions/bradc/foundations/blockdist1.chpl
  arrays/deitz/jacobi5.chpl

into nested scopes so that they take precedence over the standard
functions.

Made it so that all ambiguous functions print out in the non-developer
error message not just functions in user code.  This avoids the case
where we say a function is ambiguous but only print out the candidates
in user code, a confusing situation where only one candidate could be
printed out.

Made it so that standard module file paths are only printed out from
the modules directory.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10119 3a8e244f-b0f2-452b-bcba-4c88e055c3ca