Fixed a bug with instantiating a class with a sync type. Instantiation with variables of sync type removes the synchronization part of the type, so that the sync can be read during the constructor call, and this was happening with sync types as well, leading to a resolution error with reading a type.  Added a test case.
Added a use of the Time module to Wayne's sync array test, but this
still fails.  The problem with arrays of syncs seems to be that the
assignment to an element of an array requires a variable of the
element type (sync) thereby not allowing an assignment of a nonsync
and leading to a write.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10164 3a8e244f-b0f2-452b-bcba-4c88e055c3ca