First stab at CHANGES file.  Don't bother reading it, I'm just archiving to not lose my work before the next pass.
This draft is more or less a big laundry list of changes
sorted into rough categories.  More work needs to be done
to order the items, make sure they're in the right
categories, call out higher-level/bigger changes at the
top of the file and lump minutiae at the bottom, etc.  For
now I erred on the side of listing most major changes to
avoid additional sweeps of the CVSLOG file.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10167 3a8e244f-b0f2-452b-bcba-4c88e055c3ca