Added some more performance tests:
- ra.chpl from the release/examples/hpcc06/ directory
- cg-sparse.chpl from the npb/cg/bradc/ directory
- also added two new performance tests that look at the speed of
  compiler passes using fft.chpl from release/examples/hpcc06
  and cg-sparse.chpl as referenced above.

Fixed a bug in the performance key searcher in which
duplicate keys were being found more than once.  Now
they're found just once and the first is taken.  This
is a slight issue for the compiler passes that are
called multiple times and are currently indistinguishable.
I'm ignoring this for now (it would be nice to have the
compiler number repeats of passes, either manually or
automatically).

Updated some .cvsignore files.

Checking in my note as to where I stopped in the CVSLOG
while working on the CHANGES file.  I added this right
after checking in last week.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10174 3a8e244f-b0f2-452b-bcba-4c88e055c3ca