Template for creating a chpl man page using the txt2man utility.  Available at ~bradc/bin/ or directly from the web at: http://mvertes.free.fr/
The man page for txt2man is in ~bradc/man/man1/
Or, read the text source for the man page in the
source of the script itself, which also serves
as a nice example to base your own man page from.

chpl.txt is a stub for a Chapel source man page.

The Makefile here will build chpl.1 (the man
page) as well as chpl.pdf and chpl.html -- PDF
and HTML versions of the man page.  grabVersion.pl
is a script that reads the compiler version number
out of the compiler sources to embed into the man
page.

I'm not planning on doing more with this for now,
if anyone else wants to play with it a bit.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10176 3a8e244f-b0f2-452b-bcba-4c88e055c3ca