Oops, committed a thinko and checked in my compiler performance tests with the same name as the benchmark performance tests which caused them to share a performance data output file in the nightly regressions.  Renamed all the files and added a .perfexecopts file to the fft test so that it wouldn't run forever.
If this "multiple performance tests share the same name" continues
to be a problem, we can add a hook to name the performance data
file for a particular test.  I suspect it will be rare in practice,
though since most "real" tests tend to have a unique name and we
aren't likely to do lots of compiler + benchmark performance tests.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10186 3a8e244f-b0f2-452b-bcba-4c88e055c3ca