Changed the way that COMPSTDIN is handled by the testing system in two ways:
(1) made it so that COMPSTDIN contains the input
to be piped in rather than redirecting to a separate
file (why did I ever do that?!?);

(2) piped it into the compiler using:
      cat COMPSTDIN | chpl ...
    rather than piping it back into the compiler using:
      chpl ... < `cat COMPSTDIN`

This fixes a regression on the Mac in which its
copy of gdb complains about certain ways of piping
in stdin, and should also be more portable across
versions of gdb for the purposes of testing.

Had the PREDIFF script pipe out one more line that
gdb on the Mac prints out during debugging.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10328 3a8e244f-b0f2-452b-bcba-4c88e055c3ca