Macintoshes are case-insensitive when it comes to filenames, so when we have two modules with the same name but different cases, it causes the same file to get overwritten twice on the Macintosh. For example, a program containing a user module "sort" and the standard module "Sort" would cause the second file to overwrite the first.
This change uniquifies our output filenames when
they would be the same in a case-insensitive
OS by concatenating integers to the filename
until a unique filename is found.  Thus, in the
example above, one file would be generated with
the name sort/Sort and the second with the name
Sort2/sort2 (depending on which was generated
first).

Haven't tested this on the Mac yet, only on Linux.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10350 3a8e244f-b0f2-452b-bcba-4c88e055c3ca