Renamed a bunch of tests and updated the various README, Makefile, etc. files to reflect the changes.  All tests are now named consistently:
    mixedCaseName[-variation].chpl

where mixedCaseName describes the test succinctly and
-variation is used to call out different variations
of the test.  (for example, if we introduce other
approaches for jacobi using an array of offsets or
sparse array of offsets, those could be variations on
jacobis).

Prepended the FEATURES file with a list of tests and the
main Chapel features that they demonstrate.

Did some massaging of comments and code, but mostly this
was meant to be related to the file renaming.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10409 3a8e244f-b0f2-452b-bcba-4c88e055c3ca