Update the format of the --help message.  The descriptions now all start in the same column, starting with a newline if necessary.  Also, added an argumentOptions field to the ArgumentDescription struct to hold the text of any options that go with the argument.  Also, within each category of options, sort the options alphebetically by their long names.  I haven't done nice word wrapping yet.  The format of each option now looks like, for example:       --explain-call <call>[:<module>][:<line>]                         Explain resolution of call
Update countMemory compare files for chap07 and chap08.  Both machines are
reporting about 3 MB more memory than was in their compare files.

Fix two problems in Brad's localConfig test so it does what was intended -
removed an undefined use of "zzz" symbol, and fixed the spelling of "function"
in the .good.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10437 3a8e244f-b0f2-452b-bcba-4c88e055c3ca