Implemented idiomatic sugar coforall with a stack of sync variables and a couple small helper functions in the base module and a little bit of compiler generated code during parsing.  This is not a particularly scalable or speedy implementation but it is correct.
Added three tests.  Test 1 shows what the sugar looks like when it is
all stamped out.  Test 2 shows the sugar for test 1.  Test 3 makes
sure that the iterations are run in parallel (or at least not in
forward order).

I didn't use an array of syncs because I don't believe resizing the
array will work.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10502 3a8e244f-b0f2-452b-bcba-4c88e055c3ca