Removed "iterator" keyword from the implementation and the spec.  An iterator is now defined as a function that contains a yield statement. A function's class tag is changed to FN_ITERATOR in cleanup if the function contains a yield statement.
Changed the name of the default iterator from "ault" (as in def ault)
to "these" as suggested by Brad.  Though I think he may have been
joking at the time, "these" for iterators nicely parallels "this" for
functions.  I'm open to changing the name but I think this is the best
suggested name so far.  Added a subsection to the spec on the "these"
method.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10515 3a8e244f-b0f2-452b-bcba-4c88e055c3ca