Added checking to ensure that a non-lvalue expression is not used as an lvalue.  With the setter->ref change, non-var functions could be used on the left-hand side of an assignment statement with no effect. It is now an error.  Same for non-var iterators.
Added Symbol::isExprTemp to flag temporaries that store expressions
and thus are not valid lvalues.  This flag is turned to false during
function resolution if the temp is used to capture var functions, var
iterators, or functions with the "valid lvalue" pragma.  The "valid
lvalue" pragma is used for array and domain wrapper routines that
result in valid lvalues even though they are not var functions (since
arrays are references).

Added two test cases to check for these error messages.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10540 3a8e244f-b0f2-452b-bcba-4c88e055c3ca