Added checking to ensure that a constant value is not return by a var function.  Removed a .future.
Added a test to check that an error is generated if a var function
returns an expression that is not a valid lvalue.

Added a new future to show an issue if a var function tries to return
a slice of a global array.  This should be allowed but is not.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@10547 3a8e244f-b0f2-452b-bcba-4c88e055c3ca